requirements.txt

python-vlc
opencv-python
mediapipe
protobuf

